API consumption, invoice fragment
