+++ buyer/src/main/java/com/example/factsafrica/ui/InvoiceFragment.java
@@ -1,0 +1,39 @@
+package com.example.factsafrica.ui;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.factsafrica.R;
+import com.example.factsafrica.ui.adapter.InvoiceAdapter;
+import com.example.factsafrica.ui.models.Invoice;
+import com.example.factsafrica.ui.network.FactsAfricaApi;
+import com.example.factsafrica.ui.network.FactsAfricaClient;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import butterknife.BindView;
+import butterknife.ButterKnife;
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
+
+public class InvoiceFragment extends Fragment {
+
+    View view;
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        view = inflater.inflate(R.layout.fragment_invoice, container, false);
+        ButterKnife.bind(this, view);
+        return view;
+    }
+
+}
