+++ buyer/src/main/java/com/example/factsafrica/ui/MainActivity.java
@@ -1,8 +1,6 @@
 package com.example.factsafrica.ui;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -10,7 +8,6 @@ import android.view.View;
 import android.widget.TextView;
 
 import com.example.factsafrica.R;
-import com.example.factsafrica.ui.utils.InvoiceFragment;
 import com.google.firebase.auth.FirebaseAuth;
 
 import butterknife.BindView;
