--- buyer/src/main/java/com/example/factsafrica/ui/utils/InvoiceFragment.java
@@ -1,64 +1,0 @@
-package com.example.factsafrica.ui.utils;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.example.factsafrica.R;
-import com.example.factsafrica.ui.adapter.InvoiceAdapter;
-import com.example.factsafrica.ui.models.Invoice;
-import com.example.factsafrica.ui.network.FactsAfricaApi;
-import com.example.factsafrica.ui.network.FactsAfricaClient;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class InvoiceFragment extends Fragment {
-    @BindView(R.id.invoicesRecycler) RecyclerView mInvoicesRecycler;
-    private List<Invoice> invoices;
-    View view;
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        view = inflater.inflate(R.layout.fragment_invoice, container, false);
-        ButterKnife.bind(this, view);
-        getInvoices();
-        return view;
-    }
-
-    public void getInvoices() {
-
-        FactsAfricaApi service = FactsAfricaClient.getClient().create(FactsAfricaApi.class);
-        Call<List<Invoice>> call = service.getAllInvoices();
-        Log.v("URL", call.request().url().toString());
-        call.enqueue(new Callback<List<Invoice>>() {
-            @Override
-            public void onResponse(Call<List<Invoice>> call, Response<List<Invoice>> response) {
-                invoices = response.body();
-                InvoiceAdapter adapter = new InvoiceAdapter(invoices, view.getContext());
-                RecyclerView.LayoutManager layoutManager = new LinearLayoutManager(view.getContext(), LinearLayoutManager.VERTICAL, false);
-                mInvoicesRecycler.setLayoutManager(layoutManager);
-                mInvoicesRecycler.setHasFixedSize(true);
-                mInvoicesRecycler.setAdapter(adapter);
-                adapter.notifyDataSetChanged();
-            }
-
-            @Override
-            public void onFailure(Call<List<Invoice>> call, Throwable t) {
-
-            }
-        });
-    }
-}
