+++ buyer/src/main/res/layout/fragment_invoice.xml
@@ -5,17 +5,5 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <androidx.appcompat.widget.SearchView
-        android:id="@+id/invoiceSearch"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="5dp"
 
-        />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/invoicesRecycler"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="5dp"/>
 </LinearLayout>
